philosophia/requirements.txt
# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
